+++ gfx/drivers/gl_shaders/modern_alpha_blend.glsl.frag.h
@@ -2,9 +2,13 @@
 
 static const char *stock_fragment_modern_blend = GLSL(
    uniform sampler2D Texture;
+   uniform vec4 bgcolor;
    varying vec2 tex_coord;
    varying vec4 color;
    void main() {
-      gl_FragColor = color * texture2D(Texture, tex_coord);
+      if (bgcolor.a > 0.0)
+         gl_FragColor = bgcolor;
+      else
+         gl_FragColor = color * texture2D(Texture, tex_coord);
    }
 );